CANS-946 - Use Component Library / Design System layout (#525)
* CANS-946 - Use Component Library / Design System layout (#510)

* Always render DesignSystemLayout

* Remove ReactWoodDuckLayout

* Update header selectors for new layout

* Tune autoscroll

* Make page alerts sticky

This adjusts global page alerts to be sticky again. Browser testing
is needed. In the future, these alerts will live in a Component
Library slot on the PageLayout. For now, that isn't quite ready, so
we put them in the Body directly.

* Fine-tuned scroll adjustments

* Fix tests for page alerts

* Sticker polyfill for page alerts

* Adjust header button POM

* Move print button POM to global section
