+++ app/javascript/Application/components/Assessment/style.sass
@@ -379,7 +379,7 @@ button.change-log-print-button
 
 .top-alert-container
   position: sticky
-  top: 12rem
+  top: 14.6rem
   z-index: 1000
 
 .assessment-expand-control
