+++ app/javascript/Application/util/features.js
@@ -4,5 +4,3 @@ const activeFeatures = () => {
 export const isFeatureActive = feature => activeFeatures().indexOf(feature) >= 0
 
 export const isCans2 = () => isFeatureActive('cans2')
-
-export const isDesignSystemLayoutEnabled = () => isCans2() && isFeatureActive('enable_design_system_layout')
